# openapi/drama-watch.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Drama Watch API",
    "version": "1.0.0",
    "description": "Public request and response contract generated from the NestJS API types."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/lists/{listId}": {
      "get": {
        "operationId": "getSharedList",
        "tags": [
          "Shared lists"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedListDetailsResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "operationId": "updateSharedList",
        "tags": [
          "Shared lists"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSharedListDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedListDetailsResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "operationId": "deleteSharedList",
        "tags": [
          "Shared lists"
        ],
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/lists/{listId}/items": {
      "post": {
        "operationId": "addSharedListItem",
        "tags": [
          "Shared lists"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSharedListItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedListItemResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/public/lists/{publicSlug}": {
      "get": {
        "operationId": "getPublicSharedList",
        "tags": [
          "Shared lists"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSharedListDetailsResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "publicSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/tier-lists": {
      "get": {
        "operationId": "listTierLists",
        "tags": [
          "Tier lists"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TierListSummaryResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createTierList",
        "tags": [
          "Tier lists"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTierListDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TierListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tier-lists/{tierListId}": {
      "get": {
        "operationId": "getTierList",
        "tags": [
          "Tier lists"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TierListResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "tierListId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "operationId": "updateTierList",
        "tags": [
          "Tier lists"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTierListDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TierListResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "tierListId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "operationId": "deleteTierList",
        "tags": [
          "Tier lists"
        ],
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "parameters": [
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "tierListId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/tier-lists/{tierListId}/layout": {
      "patch": {
        "operationId": "updateTierLayout",
        "tags": [
          "Tier lists"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTierLayoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TierListResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "tierListId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/tier-lists/{tierListId}/items": {
      "post": {
        "operationId": "addTierItems",
        "tags": [
          "Tier lists"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTierItemsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TierListResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "tierListId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/tier-lists/{tierListId}/remove-item": {
      "post": {
        "operationId": "removeTierItem",
        "tags": [
          "Tier lists"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveTierItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TierListResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "tierListId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/tier-lists/{tierListId}/duplicate": {
      "post": {
        "operationId": "duplicateTierList",
        "tags": [
          "Tier lists"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TierListResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "tierListId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/public/tier-lists/{publicSlug}": {
      "get": {
        "operationId": "getPublicTierList",
        "tags": [
          "Tier lists"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicTierListResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "publicSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/public/tier-lists/share/{publicSlug}": {
      "get": {
        "operationId": "shareTierList",
        "tags": [
          "Tier lists"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Open Graph share document",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "publicSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
...
      "MediaDetails": {
        "type": "object",
        "properties": {
          "backdropPath": {
            "type": "string"
          },
          "backdropUrl": {
            "type": "string"
          },
          "posterPath": {
            "type": "string"
          },
          "posterUrl": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "tmdbId": {
            "type": "number"
          },
          "mediaType": {
            "$ref": "#/components/schemas/MediaType"
          },
          "title": {
            "type": "string"
          },
          "originalTitle": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "releaseDate": {
            "type": "string"
          },
          "firstAirDate": {
            "type": "string"
          },
          "originCountry": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "originalLanguage": {
            "type": "string"
          },
          "genreIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "tmdbVoteAverage": {
            "type": "number"
          },
          "tmdbVoteCount": {
            "type": "number"
          },
          "releaseStatus": {
            "$ref": "#/components/schemas/MediaReleaseStatus"
          },
          "runtimeMinutes": {
            "type": "number"
          },
          "totalEpisodes": {
            "type": "number"
          },
          "totalSeasons": {
            "type": "number"
          },
          "seasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaSeason"
            }
          }
        },
        "additionalProperties": false,
        "required": [
          "genreIds",
          "id",
          "mediaType",
          "originCountry",
          "originalTitle",
          "title",
          "tmdbId"
        ]
      },
...
      "PublicUserProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "displayUsername": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "joinedAt": {
            "type": "string"
          },
          "image": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "displayUsername",
          "name",
          "joinedAt"
        ],
        "additionalProperties": false
      },
...
      "SharedListVisibility": {
        "type": "string",
        "enum": [
          "private",
          "unlisted",
          "public"
        ]
      },
      "SharedListRole": {
        "type": "string",
        "enum": [
          "owner",
          "editor",
          "commenter",
          "viewer"
        ]
      },
      "SharedListDetailsResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/SharedListVisibility"
          },
          "publicSlug": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/SharedListRole"
          },
          "itemCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SharedListMemberResponse"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SharedListItemResponse"
            }
          }
        },
        "required": [
          "createdAt",
          "id",
          "itemCount",
          "items",
          "members",
          "role",
          "title",
          "updatedAt",
          "visibility"
        ],
        "additionalProperties": false
      },
      "SharedListMemberResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/PublicUserProfileResponse"
          },
          "role": {
            "$ref": "#/components/schemas/SharedListRole"
          },
          "joinedAt": {
            "type": "string"
          }
        },
        "required": [
          "user",
          "role",
          "joinedAt"
        ],
        "additionalProperties": false
      },
      "SharedListItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "mediaId": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "media": {
            "$ref": "#/components/schemas/MediaDetails"
          },
          "addedBy": {
            "$ref": "#/components/schemas/PublicUserProfileResponse"
          },
          "note": {
            "type": "string"
          },
          "groupStatus": {
            "$ref": "#/components/schemas/SharedListItemStatus"
          },
          "groupProgress": {
            "$ref": "#/components/schemas/SharedListProgressResponse"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "mediaId",
          "position",
          "media",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "SharedListItemStatus": {
        "type": "string",
        "enum": [
          "planned",
          "watching",
          "finished"
        ]
      },
      "SharedListProgressResponse": {
        "type": "object",
        "properties": {
          "currentSeason": {
            "type": "number"
          },
          "currentEpisode": {
            "type": "number"
          }
        },
        "required": [
          "currentSeason",
          "currentEpisode"
        ],
        "additionalProperties": false
      },
...
      "PublicSharedListDetailsResponse": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "unlisted",
              "public"
            ]
          },
          "publicSlug": {
            "type": "string"
          },
          "itemCount": {
            "type": "number"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SharedListMemberResponse"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicSharedListItemResponse"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "visibility",
          "publicSlug",
          "itemCount",
          "members",
          "items",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "PublicSharedListItemResponse": {
        "type": "object",
        "properties": {
          "position": {
            "type": "number"
          },
          "media": {
            "type": "object",
            "properties": {
              "releaseStatus": {
                "$ref": "#/components/schemas/MediaReleaseStatus"
              },
              "runtimeMinutes": {
                "type": "number"
              },
              "totalEpisodes": {
                "type": "number"
              },
              "totalSeasons": {
                "type": "number"
              },
              "seasons": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MediaSeason"
                }
              },
              "tmdbId": {
                "type": "number"
              },
              "mediaType": {
                "$ref": "#/components/schemas/MediaType"
              },
              "title": {
                "type": "string"
              },
              "originalTitle": {
                "type": "string"
              },
              "overview": {
                "type": "string"
              },
              "releaseDate": {
                "type": "string"
              },
              "firstAirDate": {
                "type": "string"
              },
              "originCountry": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "originalLanguage": {
                "type": "string"
              },
              "genreIds": {
                "type": "array",
                "items": {
                  "type": "number"
                }
              },
              "tmdbVoteAverage": {
                "type": "number"
              },
              "tmdbVoteCount": {
                "type": "number"
              },
              "backdropPath": {
                "type": "string"
              },
              "backdropUrl": {
                "type": "string"
              },
              "posterPath": {
                "type": "string"
              },
              "posterUrl": {
                "type": "string"
              }
            },
            "required": [
              "tmdbId",
              "mediaType",
              "title",
              "originalTitle",
              "originCountry",
              "genreIds"
            ],
            "additionalProperties": false
          },
          "addedBy": {
            "$ref": "#/components/schemas/PublicUserProfileResponse"
          },
          "note": {
            "type": "string"
          },
          "groupStatus": {
            "$ref": "#/components/schemas/SharedListItemStatus"
          },
          "groupProgress": {
            "$ref": "#/components/schemas/SharedListProgressResponse"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "position",
          "media",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
      "UpdateSharedListDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "visibility": {
            "$ref": "#/components/schemas/SharedListVisibility"
          }
        },
        "additionalProperties": false
      },
...
      "AddSharedListItemDto": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "groupStatus": {
            "$ref": "#/components/schemas/SharedListItemStatus"
          },
          "groupProgress": {
            "$ref": "#/components/schemas/SharedListProgressDto"
          }
        },
        "required": [
          "mediaId"
        ],
        "additionalProperties": false
      },
      "SharedListProgressDto": {
        "type": "object",
        "properties": {
          "currentSeason": {
            "type": "number"
          },
          "currentEpisode": {
            "type": "number"
          }
        },
        "required": [
          "currentSeason",
          "currentEpisode"
        ],
        "additionalProperties": false
      },
...
      "TierListResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/TierListVisibility"
          },
          "publicSlug": {
            "type": "string"
          },
          "revision": {
            "type": "number"
          },
          "itemCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "tiers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TierRowResponse"
            }
          },
          "unranked": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TierListMedia"
            }
          }
        },
        "required": [
          "createdAt",
          "description",
          "id",
          "itemCount",
          "revision",
          "tiers",
          "title",
          "unranked",
          "updatedAt",
          "visibility"
        ],
        "additionalProperties": false
      },
      "TierListVisibility": {
        "type": "string",
        "enum": [
          "private",
          "unlisted",
          "public"
        ]
      },
...
      "TierListSummaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/TierListVisibility"
          },
          "publicSlug": {
            "type": "string"
          },
          "revision": {
            "type": "number"
          },
          "itemCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "visibility",
          "revision",
          "itemCount",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "PublicTierListResponse": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "unlisted"
            ]
          },
          "publicSlug": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/PublicUserProfileResponse"
          },
          "tiers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string"
                },
                "color": {
                  "$ref": "#/components/schemas/TierColor"
                },
                "items": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TierListMedia"
                  }
                }
              },
              "required": [
                "label",
                "color",
                "items"
              ],
              "additionalProperties": false
            }
          },
          "itemCount": {
            "type": "number"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "description",
          "visibility",
          "publicSlug",
          "tiers",
          "itemCount",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "CreateTierListDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "additionalProperties": false
      },
      "UpdateTierListDto": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/TierListVisibility"
          }
        },
        "required": [
          "description",
          "revision",
          "title",
          "visibility"
        ],
        "additionalProperties": false
      },
      "UpdateTierLayoutDto": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "number"
          },
          "tiers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TierPlacementDto"
            }
          },
          "unrankedMediaIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "revision",
          "tiers",
          "unrankedMediaIds"
        ],
        "additionalProperties": false
      },
...
      "AddTierItemsDto": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TierMediaIdentityDto"
            }
          }
        },
        "required": [
          "items",
          "revision"
        ],
        "additionalProperties": false
      },
...
      "RemoveTierItemDto": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "number"
          },
          "mediaId": {
            "type": "string"
          }
        },
        "required": [
          "mediaId",
          "revision"
        ],
        "additionalProperties": false
      },
      "RevisionDto": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "number"
          }
        },
        "required": [
          "revision"
        ],
        "additionalProperties": false
      }
    }
  }
}